+++ CHANGELOG.md
@@ -4,12 +4,14 @@
 
 ### Added
 
+- `max_tasks` key limiting route size at vehicle level (#421)
 - Github Actions CI (#436)
 - Check for libvroom example build in CI (#514)
 
 ### Changed
 
 - CI builds now run on Ubuntu 20.04 (#455)
+- Simplified time window handling logic in TWRoute (#557)
 
 ### Removed